requirements.txt
angr==9.2.6
capstone==5.0.1
coloredlogs==6.0
Js2Py==0.74
networkx==2.5.1
numpy==1.19.4
pydot==1.4.2
pyparsing==3.0.6
python_magic==0.4.27
PyYAML==6.0.1
requests==2.27.1